Replicated style Moved tests simplified tests.
